backend/requirements.txt
fastapi
uvicorn
nltk
python-multipart
python-docx
PyMuPDF
docx2txt
google-generativeai
